refactor: 바뀐 Repository 적용
