+++ app/src/main/java/com/juniori/puzzle/domain/usecase/FetchMyFirstVideosUseCase.kt
@@ -5,6 +5,6 @@ import javax.inject.Inject
 
 class FetchMyFirstVideosUseCase @Inject constructor(private val videoRepository: Repositoryk) {
     suspend operator fun invoke(uid: String?, query: String) {
-        videoRepository.fetchMyData(uid, query)
+        videoRepository.fetchMyFirstPage(uid, query)
     }
 }
