+++ app/src/main/java/com/juniori/puzzle/domain/usecase/GetMyVideoFetchingStateUseCase.kt
@@ -1,7 +1,7 @@
 package com.juniori.puzzle.domain.usecase
 
 import com.juniori.puzzle.ui.othersgallery.Repositoryk
-import com.juniori.puzzle.ui.othersgallery.VideoFetchingState
+import com.juniori.puzzle.util.VideoFetchingState
 import kotlinx.coroutines.flow.StateFlow
 import javax.inject.Inject
