+++ app/src/main/java/com/juniori/puzzle/ui/othersgallery/OthersGalleryFragment.kt
@@ -1,16 +1,21 @@
 package com.juniori.puzzle.ui.othersgallery
 
+import android.app.Activity
 import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.activity.result.ActivityResultLauncher
+import android.widget.AdapterView
+import android.widget.ArrayAdapter
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.widget.SearchView
 import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.lifecycleScope
+import androidx.lifecycle.repeatOnLifecycle
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.google.android.material.snackbar.Snackbar
@@ -18,11 +23,12 @@ import com.juniori.puzzle.R
 import com.juniori.puzzle.databinding.FragmentOthersgalleryBinding
 import com.juniori.puzzle.ui.playvideo.PlayVideoActivity
 import com.juniori.puzzle.ui.playvideo.PlayVideoActivity.Companion.CLICKED_VIDEO_INDEX_KEY
-import com.juniori.puzzle.ui.playvideo.PlayVideoActivity.Companion.LAST_VIEWED_VIDEO_INDEX_KEY
-import com.juniori.puzzle.util.GalleryState
 import com.juniori.puzzle.util.PuzzleDialog
 import com.juniori.puzzle.util.SortType
+import com.juniori.puzzle.util.VideoFetchingState
 import dagger.hilt.android.AndroidEntryPoint
+import kotlinx.coroutines.flow.collectLatest
+import kotlinx.coroutines.launch
 
 @AndroidEntryPoint
 class OthersGalleryFragment : Fragment() {
@@ -31,10 +37,10 @@ class OthersGalleryFragment : Fragment() {
     private val binding get() = requireNotNull(_binding)
     private val viewModel: OthersGalleryViewModel by viewModels()
     private val recyclerAdapter: OtherGalleryAdapter by lazy {
-        OtherGalleryAdapter(viewModel) { clickedIndex ->
+        OthersGalleryAdapter(viewModel) { clickedIndex ->
             playVideoActivityLauncher.launch(
                 Intent(requireContext(), PlayVideoActivity::class.java).apply {
-                    this.putExtra(CLICKED_VIDEO_INDEX_KEY, clickedIndex)
+                    putExtra(CLICKED_VIDEO_INDEX_KEY, clickedIndex)
                 }
             )
         }
@@ -43,7 +49,7 @@ class OthersGalleryFragment : Fragment() {
         registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { result ->
             if (result.resultCode == Activity.RESULT_OK) {
                 val lastViewedPosition = result.data?.extras?.getInt(
-                    LAST_VIEWED_VIDEO_INDEX_KEY,
+                    PlayVideoActivity.LAST_VIEWED_VIDEO_INDEX_KEY,
                     0
                 ) ?: return@registerForActivityResult
                 binding.recycleOtherGallery.scrollToPosition(lastViewedPosition)
@@ -80,73 +86,72 @@ class OthersGalleryFragment : Fragment() {
         }
 
         binding.otherGallerySwipeRefresh.setOnRefreshListener {
-            viewModel.getMainData()
+            viewModel.fetchFirstVideoPage()
         }
 
-        viewModel.list.observe(viewLifecycleOwner) { dataList ->
-            binding.otherGallerySwipeRefresh.isRefreshing = false
-
-            recyclerAdapter.submitList(dataList)
-
-            binding.textOtherGalleryNotFound.isVisible = dataList.isEmpty()
-        }
-
-        viewModel.refresh.observe(viewLifecycleOwner) { isRefresh ->
-            binding.progressOtherGallery.isVisible = isRefresh
-        }
-
-        viewModel.state.observe(viewLifecycleOwner) { state ->
-            when (state) {
-                GalleryState.NONE -> {
-                    snackBar?.dismiss()
-                }
-
-                GalleryState.END_PAGING -> {
-                    snackBar = Snackbar.make(
-                        view,
-                        R.string.gallery_end_paging,
-                        Snackbar.LENGTH_SHORT
-                    ).apply {
-                        setAction(R.string.gallery_check) {
-                            dismiss()
-                        }
-                    }
-
-                    snackBar?.show()
-                }
-
-                GalleryState.NETWORK_ERROR_PAGING -> {
-                    snackBar =
-                        Snackbar.make(
-                            view,
-                            R.string.gallery_paging_error,
-                            Snackbar.LENGTH_INDEFINITE
-                        )
-                            .setAction(R.string.gallery_retry) {
-                                viewModel.getPaging()
-                            }
-                    snackBar?.show()
-                }
-
-                GalleryState.NETWORK_ERROR_BASE -> {
+        viewLifecycleOwner.lifecycleScope.launch {
+            repeatOnLifecycle(Lifecycle.State.STARTED) {
+                viewModel.videoList.collectLatest { videoList ->
                     binding.otherGallerySwipeRefresh.isRefreshing = false
-                    snackBar =
-                        Snackbar.make(
-                            view,
-                            R.string.gallery_init_load_error,
-                            Snackbar.LENGTH_INDEFINITE
-                        )
-                            .setAction(R.string.gallery_retry) {
-                                viewModel.getMainData()
-                            }
-                    snackBar?.show()
+
+                    recyclerAdapter.submitList(videoList)
+
+                    binding.textOtherGalleryNotFound.isVisible = videoList.isEmpty()
                 }
             }
         }
 
-        viewModel.list.value.also { list ->
-            if (list == null || list.isEmpty()) {
-                viewModel.getMainData()
+        viewLifecycleOwner.lifecycleScope.launch {
+            repeatOnLifecycle(Lifecycle.State.STARTED) {
+                viewModel.videoFetchingState.collectLatest { state ->
+                    binding.progressOtherGallery.isVisible = false
+
+                    when (state) {
+                        VideoFetchingState.NONE -> {
+                            snackBar?.dismiss()
+                        }
+
+                        VideoFetchingState.Loading -> {
+                            binding.progressOtherGallery.isVisible = true
+                        }
+
+                        VideoFetchingState.NO_MORE_VIDEO -> {
+                            snackBar = Snackbar.make(
+                                view,
+                                R.string.gallery_end_paging,
+                                Snackbar.LENGTH_SHORT
+                            ).apply {
+                                setAction(R.string.gallery_check) {
+                                    dismiss()
+                                }
+                            }
+                            snackBar?.show()
+                        }
+
+                        VideoFetchingState.NETWORK_ERROR_PAGING -> {
+                            snackBar = Snackbar.make(
+                                view,
+                                R.string.gallery_paging_error,
+                                Snackbar.LENGTH_INDEFINITE
+                            ).setAction(R.string.gallery_retry) {
+                                viewModel.fetchNextVideoPage()
+                            }
+                            snackBar?.show()
+                        }
+
+                        VideoFetchingState.NETWORK_ERROR_BASE -> {
+                            binding.otherGallerySwipeRefresh.isRefreshing = false
+                            snackBar = Snackbar.make(
+                                view,
+                                R.string.gallery_init_load_error,
+                                Snackbar.LENGTH_INDEFINITE
+                            ).setAction(R.string.gallery_retry) {
+                                viewModel.fetchFirstVideoPage()
+                            }
+                            snackBar?.show()
+                        }
+                    }
+                }
             }
         }
 
@@ -175,7 +180,8 @@ class OthersGalleryFragment : Fragment() {
                 }
             }
 
-                override fun onNothingSelected(parent: AdapterView<*>?) {}
+                override fun onNothingSelected(parent: AdapterView<*>?) {
+                }
             }
             popup.dismissPopupList()
         }
@@ -207,6 +213,5 @@ class OthersGalleryFragment : Fragment() {
 
     companion object {
         const val RECYCLER_TOP = 0
-        const val VIDEO_EXTRA_NAME = "videoInfo"
     }
 }
