+++ app/src/main/java/com/juniori/puzzle/util/VideoFetchingState.kt
@@ -1,0 +1,9 @@
+package com.juniori.puzzle.util
+
+enum class VideoFetchingState {
+    NONE,
+    Loading,
+    NO_MORE_VIDEO,
+    NETWORK_ERROR_PAGING,
+    NETWORK_ERROR_BASE,
+}
