+++ app/src/main/res/navigation/main_bottomnavigation.xml
@@ -30,7 +30,7 @@
 
     <fragment
         android:id="@+id/fragment_main_othersgallery"
-        android:name="com.juniori.puzzle.ui.othersgallery.OthersGalleryFragmentk"
+        android:name="com.juniori.puzzle.ui.othersgallery.OthersGalleryFragment"
         android:label="@string/main_othersgallery"
         tools:layout="@layout/fragment_othersgallery" />
